Updated userManagement
